+++ ServerModule/app/src/users.py
@@ -1,10 +1,9 @@
 from abc import ABC, abstractmethod
 
 from src.plants import Plant
-from src.devices import Device, create_device_from_type
+from src.devices import create_device_from_type
 from src.db.db_utils import DBInterface
 from src.logger import Logger
-from src.measurements import Moisture
 from src.thread_manager import PlantThreadManager
 from src.textbook import Textbook
 
@@ -75,7 +74,7 @@ class Consumer(User):
         req_brightness: float,
         req_humidity: float,
         req_temperature: float,
-        req_moisture: int,
+        req_moisture: float,
         description: str = None,
         care_instructions: str = None,
         location: str | None = None,
